Extract Locale to a separate file and add tests
